feat: define `FuncV` dialect and full denotation
`FuncV` is a dialect for functions-as-(SSA)-values. The lowering of `FuncV` will capture the step of inlining. This way, a complex lowering that maps expressions to large inlined computations can be turned into an expression-to-expression map by first lowering to `FuncV` and bundling those large computaitons into functions at the beginning of the lowered program.